+++ client/src/App.js
@@ -2,9 +2,8 @@
 import { Route, Routes } from "react-router-dom";
 import About from "./pages/About";
 import Home from "./pages/Home";
-import Navbar from "./Navbar";
+import Navbar from "./components/Navbar/Navbar";
 import "./App.css";
-
 
 const App = () => (
 	<div>
