+++ client/src/components/Navbar/Navbar.css
@@ -1,0 +1,27 @@
+nav {
+    background-color: #F2EFF0;
+    width: 100%;
+}
+
+.navbar-wrapper {
+    display: flex;
+    justify-content: space-between;
+    margin-inline: auto;
+}
+
+.logo {
+    padding-left: 10px;
+    padding-top: 20px;
+    width: 10rem;
+    margin-bottom: 18px;
+}
+
+.link {
+    font-size: x-large;
+    font-weight: bold;
+}
+
+#logInOut {
+    background-color: transparent;
+    border: 0;
+}
