--- client/src/pages/Home.css
@@ -1,11 +1,0 @@
-.logo {
-	width: 100px;
-}
-
-.message {
-	color: blue;
-}
-
-.message:hover {
-	background-color: yellow;
-}
